+++ ReleaseNotes.md
@@ -9,6 +9,7 @@
 - cta/CTA#147 - Development of the postgres SchedulerDatabase
 - cta/CTA#259 - cta-rmcd should not exit if /dev/sg0 is missing
 - cta/CTA#345 - Remove some objectstore dependencies outside scheduler
+- cta/CTA#340 - Fix catalogue recompiling
 
 ### Continuous Integration
 - cta/CTA#262 - Manage CI repositories with `cta-release` code